separate custodian tx building
