+++ aggregator/aggregator-main/src/transaction/mod.rs
@@ -1,3 +1,2 @@
 mod clear;
-mod custodian;
 mod leap;
