+++ aggregator/util/common/src/lib.rs
@@ -1,4 +1,5 @@
 #![allow(missing_docs)]
 
 pub mod error;
+pub mod types;
 pub mod utils;
