+++ aggregator/util/common/src/types.rs
@@ -1,0 +1,5 @@
+pub enum RequestType {
+    CkbToBranch = 1,
+    BranchToCkb = 2,
+    BranchToBranch = 3,
+}
