+++ aggregator/util/rgbpp-tx/Cargo.toml
@@ -7,6 +7,8 @@ edition = "2021"
 
 [dependencies]
 aggregator-common = { path = "../common", version = "0.1.0" }
+ckb-app-config = { path = "../../../util/app-config", version = "= 0.116.1" }
+ckb-stop-handler = { path = "../../../util/stop-handler", version = "= 0.116.1" }
 
 ckb-gen-types = { version = "=0.116.1", default-features = false }
 ckb-logger = "=0.116.1"
